Revert "etWET"
This reverts commit 32831b96f18e90e28b353699c4937c2c4cfd0603, reversing
changes made to 99e0e55cf20eedb665621defed8bf703ceb7cc6a.

+++ source/Natural Selection Sim/Logic/Carivore.cs
@@ -2,7 +2,7 @@ namespace Natural_Selection_Sim
 {
     public class Carnivore : Entity
     {
-        public Carnivore(double b, double d, double m, double s, double z)
+        public Carnivore(float b, float d, float m, float s, float z)
             : base(b, d, m, s, z) { }
 
         public Carnivore(Entity parent)
